# backend/docs/swagger.yaml
definitions:
...
  protocol.WindowsWebAddEVCertificateReq:
    properties:
      isMicrosoftVerifyCertificate:
        description: 是用于微软网站校验的证书
        example: false
        type: boolean
      machineIp:
        description: 机器 IP
        example: 172.16.1.1
        type: string
      notAfter:
        description: 失效时间
        example: "2026-01-20 09:48:00"
        type: string
      notBefore:
        description: 生效时间
        example: "2026-01-20 09:48:00"
        type: string
      owner:
        description: 所有者
        example: C=CN, ST=HuNan, L=Changsha, O=ivfzhou, OU=ivfzhou, CN=windows personal
          ev certificate, emailAddress=ivfzhou@126.com
        type: string
      password:
        description: 密码
        example: "123456"
        maxLength: 64
        minLength: 6
        type: string
      publicKeyAlgorithm:
        description: 公钥算法
        example: SHA1WithRSA
        type: string
      publisher:
        description: 颁发者
        example: C=CN, ST=HuNan, L=Changsha, O=ivfzhou, OU=ivfzhou, CN=windows personal
          ev certificate, emailAddress=ivfzhou@126.com
        type: string
      serialNumber:
        description: 序列号
        example: 6e4314c191e2fe571b60818914982e53e03e7a20
        maxLength: 1024
        type: string
      sha1:
        description: 指纹
        example: e4fd2280d075253896580e0c061730032e2b9c7a
        type: string
      signatureAlgorithm:
        description: 签名算法
        example: SHA1WithRSA
        type: string
      type:
        description: 类型
        enum:
        - 1
        - 3
        example: 1
        type: integer
      version:
        description: 版本号
        example: 3
        type: integer
    required:
    - notAfter
    - notBefore
    - owner
    - password
    - publicKeyAlgorithm
    - publisher
    - serialNumber
    - signatureAlgorithm
    type: object
...
  util.Response-any:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/errs.Code'
        description: 响应码，大于 0 表示错误
        example: 0
      data:
        description: 响应数据
      message:
        description: 提示语
        example: 失败时的提示语
        type: string
    type: object
...
host: 127.0.0.1:8090
info:
  contact:
    email: ivfzhou@126.com
    name: ivfzhou
  license:
    name: MulanPSL-2.0 license
    url: http://license.coscl.org.cn/MulanPSL2
  title: CSMS API
  version: "1.0"
paths:
...
  /web/windows/addEVCertificate:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 请求体
        in: body
        name: _
        required: true
        schema:
          $ref: '#/definitions/protocol.WindowsWebAddEVCertificateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-any'
      summary: 添加 EV 证书
      tags:
      - Windows-WebAPI
  /web/windows/deleteCertificate/{appId}:
    delete:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 应用 ID
        in: path
        name: appId
        required: true
        type: string
      - description: 证书 ID
        example: 4ef83c03e2ce4f1f94c11168d1acd087
        in: query
        name: certificateId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-any'
      summary: 删除证书
      tags:
      - Windows-WebAPI
  /web/windows/downloadCertificate/{appId}:
    get:
      consumes:
      - application/x-www-urlencoded
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 应用 ID
        in: path
        name: appId
        required: true
        type: string
      - description: 证书 ID
        example: 93aa1dc7f4ab32bdc5a3b0b78ecc35b5
        in: query
        name: certificateId
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: 下载证书
      tags:
      - Windows-WebAPI
  /web/windows/downloadCompanyCertificate:
    get:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 证书 ID
        example: 4ef83c03e2ce4f1f94c11168d1acd087
        in: query
        name: certificateId
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: 下载公司证书
      tags:
      - Windows-WebAPI
  /web/windows/getCertificatePassword/{appId}:
    get:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 应用 ID
        in: path
        name: appId
        required: true
        type: string
      - description: 证书 ID
        example: 4ef83c03e2ce4f1f94c11168d1acd087
        in: query
        name: certificateId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-protocol_WindowsWebGetCertificatePasswordRsp'
      summary: 查看证书密码
      tags:
      - Windows-WebAPI
  /web/windows/grantAppEVCertificate:
    post:
      consumes:
      - application/json
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 请求体
        in: body
        name: _
        required: true
        schema:
          $ref: '#/definitions/protocol.WindowsWebGrantAppEVCertificateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-any'
      summary: 授权应用使用个人 EV 证书
      tags:
      - Windows-WebAPI
  /web/windows/listCertificates/{appId}:
    get:
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 应用 ID
        in: path
        name: appId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-protocol_WindowsWebListCertificatesRsp'
      summary: 证书列表
      tags:
      - Windows-WebAPI
  /web/windows/listCompanyCertificates:
    get:
      consumes:
      - application/json
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-protocol_WindowsWebListCompanyCertificatesRsp'
      summary: 获取后台管理中的 Windows 证书
      tags:
      - Windows-WebAPI
  /web/windows/listGrantCertificateApps:
    get:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 应用 ID
        example: 4ef83c03e2ce4f1f94c11168d1acd087
        in: query
        name: appId
        type: string
      - description: 证书 ID
        example: e4fd2280d075253896580e0c06173003
        in: query
        name: certificateId
        type: string
      - description: 页数
        example: 1
        in: query
        name: pageNumber
        type: integer
      - description: 每页条数
        example: 10
        in: query
        maximum: 100
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-protocol_WindowsWebListGrantCertificateAppsRsp'
      summary: 获取授权 Windows EV 证书应用列表
      tags:
      - Windows-WebAPI
  /web/windows/listSigningJobs/{appId}:
    get:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 应用 ID
        in: path
        name: appId
        required: true
        type: string
      - description: 关键字
        in: query
        name: keyWord
        type: string
      - description: 页数
        example: 1
        in: query
        minimum: 1
        name: pageNumber
        type: integer
      - description: 一页任务条数
        example: 10
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        type: integer
      - description: 任务类型
        enum:
        - 1
        - 2
        - 3
        - 4
        example: 1
        in: query
        name: signingType
        type: integer
      - description: 任务状态
        enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        - 7
        example: 1
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-protocol_WindowsWebListSigningJobsRsp'
      summary: 获取签名任务列表
      tags:
      - Windows-WebAPI
  /web/windows/listWHQLJobs/{appId}:
    get:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 应用 ID
        in: path
        name: appId
        required: true
        type: string
      - description: 页数
        example: 1
        in: path
        minimum: 1
        name: pageNumber
        type: integer
      - description: 一页任务条数
        example: 10
        in: path
        maximum: 100
        minimum: 1
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-protocol_WindowsWebListWHQLJobsRsp'
      summary: 获取 WHQL 任务列表
      tags:
      - Windows-WebAPI
  /web/windows/removeCompanyCertificate:
    delete:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: 请求日期
        example: Mon, 02 Jan 2006 15:04:05 GMT
        in: header
        name: Date
        required: true
        type: string
      - description: 会话凭据
        example: csms_user=; csms_seesion=
        in: header
        name: Cookie
        required: true
        type: string
      - description: 证书 ID
        example: 4ef83c03e2ce4f1f94c11168d1acd087
        in: path
        name: certificateId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response-any'
      summary: 删除公司证书
      tags:
      - Windows-WebAPI
...
schemes:
- http
swagger: "2.0"